% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.R
\name{apply_damage}
\alias{apply_damage}
\title{Apply a simple damage function}
\usage{
apply_damage(exposure, intensity, curve)
}
\arguments{
\item{exposure}{An exposure object}

\item{intensity}{Numeric vector of hazard intensity (same length)}

\item{curve}{A function mapping intensity → damage ratio}
}
\description{
Apply a simple damage function
}
