% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exposures.R
\name{exposure}
\alias{exposure}
\title{Create an exposure (in-force) object}
\usage{
exposure(x, id, value)
}
\arguments{
\item{x}{An sf object with POINT geometry}

\item{id}{Column name for exposure ID}

\item{value}{Column name for exposure value (e.g. TIV)}
}
\value{
An object of class `exposure`
}
\description{
Create an exposure (in-force) object
}
