% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validate_exposure}
\alias{validate_exposure}
\title{validate_exposure}
\usage{
validate_exposure(x)
}
\arguments{
\item{x}{Of Class Exposure}
}
\value{
A True/False value
}
\description{
validate_exposure
}
